Feat: 専ブラからの書き込みに対応 一部のページは普通のブラウザからも利用可能です mCaptchaは仕様が他のCaptchaと違ったため対応を打ち切ります

+++ mintMigration/versions/e5b09da894a1_create_threads_table.py
@@ -21,7 +21,7 @@ depends_on: Union[str, Sequence[str], None] = None
 def upgrade() -> None:
     op.create_table(
         "threads",
-        sa.Column("id", sa.Integer, primary_key=True),
+        sa.Column("id", sa.BigInteger, primary_key=True),
         sa.Column("board", sa.String(), nullable=False),
         sa.Column("title", sa.String(), nullable=False),
         sa.Column("name", sa.String(), nullable=False),
